Changes for adding images to mailbox signature #366
