+++ app/Http/Controllers/SecureController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\ActivityLog;
+use App\Misc\Helper;
 use App\SendLog;
 use App\Thread;
 use App\User;
@@ -176,5 +177,41 @@ class SecureController extends Controller
         }
 
         return redirect()->route('logs', ['name' => $name]);
+    }
+
+    /**
+     * Upload files and images.
+     */
+    public function upload(Request $request)
+    {
+        $response = [
+            'status' => 'error',
+            'msg'    => '', // this is error message
+        ];
+
+        $user = auth()->user();
+
+        if (!$user) {
+            $response['msg'] = __('Please login to upload file');
+        }
+
+        if (!$request->hasFile('file') || !$request->file('file')->isValid() || !$request->file) {
+            $response['msg'] = __('Error occured uploading file');
+        }
+
+        if (!$response['msg']) {
+
+            $upload = Helper::uploadFile($request->file, ['jpg','gif','png']);
+            $filename = basename($upload);
+
+            if ($upload) {
+                $response['status'] = 'success';
+                $response['url'] = Helper::uploadedFileUrl($filename);
+            } else {
+                $response['msg'] = __('Error occured uploading file');
+            }
+        }
+
+        return \Response::json($response);
     }
 }
